gmarkup: fix type of length parameter of text_validate()
A bug hunter submitted a bogus vulnerability report (#YWH-PGM9867-188)
regarding this issue. the code passes the size of a GString, gsize, to
an int, which is surely a bug. Fortunately, g_utf8_validate_len() also
accepts a gsize, so there's only one spot that needs to be changed.

+++ glib/gmarkup.c
@@ -492,7 +492,7 @@ name_validate (GMarkupParseContext  *context,
 static gboolean
 text_validate (GMarkupParseContext  *context,
                const gchar          *p,
-               gint                  len,
+               gsize                 len,
                GError              **error)
 {
   if (!g_utf8_validate_len (p, len, NULL))
